+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 <!-- CHANGELOG NEWEST -->
 - FIX RxDocument getter should return the same object on property paths [#4548](https://github.com/pubkey/rxdb/pull/4548)
 - OPTIMIZE use `findDocumentsById()` on queries that do only do an `$in` operator on the primary key.
+- FIX `$or` query with other operator on dexie.js
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->