+++ CHANGELOG.md
@@ -2,14 +2,18 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
-- FIX RxDocument getter should return the same object on property paths [#4548](https://github.com/pubkey/rxdb/pull/4548)
-- OPTIMIZE use `findDocumentsById()` on queries that do only do an `$in` operator on the primary key.
-- FIX `$or` query with other operator on dexie.js
+
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->
 
 <!-- RELEASE BELOW -->
+
+### 14.3.0 (13 March 2023)
+
+- FIX RxDocument getter should return the same object on property paths [#4548](https://github.com/pubkey/rxdb/pull/4548)
+- OPTIMIZE use `findDocumentsById()` on queries that do only do an `$in` operator on the primary key.
+- FIX `$or` query with other operator on dexie.js
 
 ### 14.2.3 (10 March 2023)